Changed input from float literal to string literal

+++ lib/long_arithmetic.h
@@ -10,7 +10,7 @@ class LongNum {
     std::deque<long long> float_part;
     int n_frac{};
     // numbers closer to the beginning is less than the ones that closer to end f_part[0] < f_part[1]
-    const int BASE = 1000000;
+    const long long BASE = 1000000;
     bool isNegative;
 
 public:
